Create separate pages for GamePage and StartPage.
