+++ OxbowCastle/GamePage.xaml
@@ -1,0 +1,46 @@
+<?xml version="1.0" encoding="utf-8"?>
+<Page
+    x:Class="OxbowCastle.GamePage"
+    xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
+    xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
+    xmlns:local="using:OxbowCastle"
+    xmlns:d="http://schemas.microsoft.com/expression/blend/2008"
+    xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006"
+    mc:Ignorable="d"
+    Background="{ThemeResource ApplicationPageBackgroundThemeBrush}">
+
+    <Grid>
+        <Grid.Resources>
+            <Style x:Key="HeadingParaStyle" x:Name="m_headingParaStyle" TargetType="TextBlock" BasedOn="{StaticResource HeadingStyle}">
+                <Setter Property="TextWrapping" Value="Wrap"/>
+                <Setter Property="Margin" Value="10,10,0,0"/>
+            </Style>
+            <Style x:Key="BodyParaStyle" x:Name="m_bodyParaStyle" TargetType="TextBlock" BasedOn="{StaticResource BodyStyle}">
+                <Setter Property="TextWrapping" Value="Wrap"/>
+                <Setter Property="Margin" Value="10,5,0,5"/>
+            </Style>
+            <Style x:Name="m_listParaStyle" TargetType="TextBlock" BasedOn="{StaticResource BodyStyle}">
+                <Setter Property="TextWrapping" Value="Wrap"/>
+                <Setter Property="Margin" Value="35,5,0,0"/>
+            </Style>
+            <Style x:Name="m_listBulletStyle" TargetType="TextBlock" BasedOn="{StaticResource BodyStyle}">
+                <Setter Property="Margin" Value="20,5,0,0"/>
+            </Style>
+            <Style x:Key="CommandStyle" x:Name="m_commandStyle" TargetType="TextBlock">
+                <Setter Property="FontSize" Value="16"/>
+                <Setter Property="FontWeight" Value="Bold"/>
+                <Setter Property="TextWrapping" Value="Wrap"/>
+                <Setter Property="Margin" Value="0,20,0,0"/>
+                <Setter Property="Foreground" Value="{StaticResource AccentTextFillColorSecondaryBrush}"/>
+            </Style>
+        </Grid.Resources>
+        <Grid.RowDefinitions>
+            <RowDefinition Height="*"/>
+            <RowDefinition Height="Auto"/>
+        </Grid.RowDefinitions>
+        <ScrollViewer x:Name="m_outputScrollViewer" Background="{StaticResource AcrylicBackgroundFillColorDefaultBrush}">
+            <StackPanel x:Name="m_outputStackPanel" Margin="20,10,20,10"/>
+        </ScrollViewer>
+        <TextBox Grid.Row="1" x:Name="m_commandTextBox" KeyDown="TextBox_KeyDown" PlaceholderText="Type a command and press ENTER."/>
+    </Grid>
+</Page>
