+++ OxbowCastle/MainWindow.xaml
@@ -8,70 +8,6 @@
     xmlns:d="http://schemas.microsoft.com/expression/blend/2008"
     xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006"
     mc:Ignorable="d">
-    
-    <Grid Background="{StaticResource AccentAAFillColorSecondaryBrush}">
-        <Grid.Resources>
-            <Style x:Key="HeadingStyle" TargetType="TextBlock">
-                <Setter Property="FontSize" Value="24"/>
-                <Setter Property="FontWeight" Value="Bold"/>
-            </Style>
-            <Style x:Key="BodyStyle" TargetType="TextBlock">
-                <Setter Property="FontSize" Value="16"/>
-            </Style>
-            <Style x:Key="HeadingParaStyle" x:Name="m_headingParaStyle" TargetType="TextBlock" BasedOn="{StaticResource HeadingStyle}">
-                <Setter Property="TextWrapping" Value="Wrap"/>
-                <Setter Property="Margin" Value="10,10,0,0"/>
-            </Style>
-            <Style x:Key="BodyParaStyle" x:Name="m_bodyParaStyle" TargetType="TextBlock" BasedOn="{StaticResource BodyStyle}">
-                <Setter Property="TextWrapping" Value="Wrap"/>
-                <Setter Property="Margin" Value="10,5,0,5"/>
-            </Style>
-            <Style x:Name="m_listParaStyle" TargetType="TextBlock" BasedOn="{StaticResource BodyStyle}">
-                <Setter Property="TextWrapping" Value="Wrap"/>
-                <Setter Property="Margin" Value="35,5,0,0"/>
-            </Style>
-            <Style x:Name="m_listBulletStyle" TargetType="TextBlock" BasedOn="{StaticResource BodyStyle}">
-                <Setter Property="Margin" Value="20,5,0,0"/>
-            </Style>
-            <Style x:Key="CommandStyle" x:Name="m_commandStyle" TargetType="TextBlock">
-                <Setter Property="FontSize" Value="16"/>
-                <Setter Property="FontWeight" Value="Bold"/>
-                <Setter Property="TextWrapping" Value="Wrap"/>
-                <Setter Property="Margin" Value="0,20,0,0"/>
-                <Setter Property="Foreground" Value="{StaticResource AccentTextFillColorSecondaryBrush}"/>
-            </Style>
-        </Grid.Resources>
-        
-        <ItemsControl x:Name="m_gameListControl" HorizontalAlignment="Center" VerticalAlignment="Center">
-            <ItemsControl.ItemsPanel>
-                <ItemsPanelTemplate>
-                    <ItemsStackPanel/>
-                </ItemsPanelTemplate>
-            </ItemsControl.ItemsPanel>
-            <ItemsControl.ItemTemplate>
-                <DataTemplate x:DataType="local:GameInfo">
-                    <Button Width="600" Margin="10" Click="GameButton_Click">
-                        <Grid>
-                            <Grid.RowDefinitions>
-                                <RowDefinition Height="Auto"/>
-                                <RowDefinition Height="Auto"/>
-                            </Grid.RowDefinitions>
-                            <TextBlock HorizontalAlignment="Center" Text="{x:Bind Action}" Style="{StaticResource BodyStyle}"/>
-                            <TextBlock Grid.Row="1" HorizontalAlignment="Center" Text="{x:Bind Name}" Style="{StaticResource HeadingStyle}"/>
-                        </Grid>
-                    </Button>
-                </DataTemplate>
-            </ItemsControl.ItemTemplate>
-        </ItemsControl>
-        <Grid x:Name="m_gameControl" Visibility="Collapsed">
-            <Grid.RowDefinitions>
-                <RowDefinition Height="*"/>
-                <RowDefinition Height="Auto"/>
-            </Grid.RowDefinitions>
-            <ScrollViewer x:Name="m_outputScrollViewer" Background="{StaticResource AcrylicBackgroundFillColorDefaultBrush}">
-                <StackPanel x:Name="m_outputStackPanel" Margin="20,10,20,10"/>
-            </ScrollViewer>
-            <TextBox Grid.Row="1" x:Name="m_commandTextBox" KeyDown="TextBox_KeyDown" PlaceholderText="Type a command and press ENTER."/>
-        </Grid>
-    </Grid>
+
+    <Frame x:Name="m_frame"/>
 </Window>
